@uppy/aws-s3: remove legacy plugin
