+++ packages/@uppy/aws-s3-multipart/package.json
@@ -23,18 +23,7 @@
     "url": "git+https://github.com/transloadit/uppy.git"
   },
   "dependencies": {
-    "@uppy/companion-client": "workspace:^",
-    "@uppy/utils": "workspace:^"
-  },
-  "devDependencies": {
-    "@aws-sdk/client-s3": "^3.362.0",
-    "@aws-sdk/s3-request-presigner": "^3.362.0",
-    "nock": "^13.1.0",
-    "vitest": "^1.2.1",
-    "whatwg-fetch": "3.6.2"
-  },
-  "peerDependencies": {
-    "@uppy/core": "workspace:^"
+    "@uppy/aws-s3": "workspace:^"
   },
   "stableVersion": "3.11.0"
 }
