+++ packages/@uppy/aws-s3/package.json
@@ -11,7 +11,8 @@
     "amazon s3",
     "s3",
     "uppy",
-    "uppy-plugin"
+    "uppy-plugin",
+    "multipart"
   ],
   "homepage": "https://uppy.io",
   "bugs": {
@@ -22,13 +23,13 @@
     "url": "git+https://github.com/transloadit/uppy.git"
   },
   "dependencies": {
-    "@uppy/aws-s3-multipart": "workspace:^",
     "@uppy/companion-client": "workspace:^",
-    "@uppy/utils": "workspace:^",
-    "@uppy/xhr-upload": "workspace:^",
-    "nanoid": "^4.0.0"
+    "@uppy/utils": "workspace:^"
   },
   "devDependencies": {
+    "@aws-sdk/client-s3": "^3.362.0",
+    "@aws-sdk/s3-request-presigner": "^3.362.0",
+    "nock": "^13.1.0",
     "vitest": "^1.2.1",
     "whatwg-fetch": "3.6.2"
   },
