+++ packages/@uppy/aws-s3/src/HTTPCommunicationQueue.ts
