+++ packages/@uppy/aws-s3/src/createSignedURL.test.ts
