+++ packages/@uppy/aws-s3/src/createSignedURL.ts
