+++ packages/@uppy/aws-s3/src/index.test.ts
