--- packages/@uppy/aws-s3/src/locale.js
@@ -1,5 +1,0 @@
-export default  {
-  strings: {
-    timedOut: 'Upload stalled for %{seconds} seconds, aborting.',
-  },
-}
