+++ packages/@uppy/aws-s3/src/utils.ts
