+++ packages/@uppy/aws-s3/tsconfig.build.json
