+++ packages/@uppy/aws-s3/tsconfig.json
